perf: fix decoder streams to make pooled connections reusable (#2484)
When a response body is being decompressed, and the length wasn't known, but was using chunked transfer-encoding, the remaining `0\r\n\r\n` was not consumed. That would leave the connection in a state that could be not be reused, and so the pool had to discard it.

This fix makes sure the remaining end chunk is consumed, improving the amount of pooled connections that can be reused.

Closes #2381